Fix isList to check recursively.
BiwaScheme.isList (also used by assert_list) only checked to see if the
obj was a pair.  It now properly checks the list recursively, ending
with the (new) empty list.  It still does not check if it's cyclic.

This is believed to cause a few test cases to fail where the test cases
were relying on the old, incorrect implementation of assert_list.